first lesson content and fixing index
