+++ src/pages/Lessons/les1/index.mdx
@@ -9,9 +9,15 @@ exerpt: "Introduction Lesson to the first programming semester"
 import Vocab from '../../../components/ui/Vocab';
 import sc from '../../../images/scratchExample.png';
 import vr from './vrm.jpg';
+import out from './out.jpg';
+import control from './control.jpg';
+
+import sense from './sense.jpg';
+import sound from './sound.jpg';
+import click from './click.jpg';
 
 
-## Activity 1 : Tech Ice Breakers
+## Activity 1 : Tech Ice Breakers [10 min]
 1. Hardware vs software
 2. What are you favorite Applications?
 3. What are you favorite Gaming Consoles?
@@ -23,7 +29,7 @@ import vr from './vrm.jpg';
 
 Activity: Draw Super Computer with all Computer Parts
 
-## Activity 2 : Scratch
+## Activity 2 : Scratch [20 min]
 
 
 <Vocab words={['Scratch', 'Motion', 'looks','Sound', 'Pen', 'Control', 'Sensing']} link="https://wwww.quizlet.com" title="Scratch Vocab"/> 
@@ -32,34 +38,32 @@ Activity: Draw Super Computer with all Computer Parts
 <img src={vr} class="test" />
 
 ### looks
-magazine pic
+<img src={out} class="test" />
 
 ### Sound
-concert
+<img src={sound} class="test" />
 
 ### Pen
-clicking
+<img src={click} class="test" />
 
 ### Control 
-traffic guy
+<img src={control} class="test" />
 
 ### Sensing
-5 senses
+<img src={sense} class="test" />
 
 Activity : First Scratch App (Say your favorite things!)
 <img src={sc} class="test"/>
 
 
 
-## Activity 3 : Customize Your Scratch
+## Activity 3 : Customize Your Scratch / Explore Scratch Projects [10 min]
 
-### costumes
+Activity Create (optional) : create different costumes for you sprite
 
-### background
-
-Activity Make your own programming
+Activity Explore (optional) : Explore some of the projects online
 
 
-## Activity 4 **Advance** : Input / Output device
+## Activity 4 **Advance** : Input / Output device [10 min]
 
 Activity : Make Scratch program determining whether the hardware is input or output
